Move auxiliary security header to its own file
