+++ src/mac/frame/security/auxiliary_security_header.rs
@@ -1,0 +1,145 @@
+//! All auxiliary security header structs and functions
+
+use byte::{BytesExt, TryRead, TryWrite, LE};
+
+use super::{KeyIdentifierMode, SecurityControl};
+
+/// A struct describing the Auxiliary Security Header
+/// See: section 7.4 of the 802.15.4-2011 standard
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+pub struct AuxiliarySecurityHeader {
+    /// The control field in the Auxiliary Security Header
+    pub control: SecurityControl,
+    /// The frame counter
+    pub frame_counter: u32,
+    /// If the key_identifier field in [`AuxiliarySecurityHeader::control`] is not set to None, this field contains the key identifier
+    /// of this frame, otherwise it is None
+    pub key_identifier: Option<KeyIdentifier>,
+}
+
+impl AuxiliarySecurityHeader {
+    /// Get the size of this security header, in octets
+    pub fn get_octet_size(&self) -> u8 {
+        // SecurityControl length + FrameCounter length
+        let length = 1
+            + 4
+            + match self.key_identifier {
+                Some(key_id) => match key_id.key_source {
+                    Some(source) => match source {
+                        KeySource::Short(_) => 5,
+                        KeySource::Long(_) => 9,
+                    },
+                    None => 1,
+                },
+                None => 0,
+            };
+        length
+    }
+}
+
+impl TryRead<'_> for AuxiliarySecurityHeader {
+    fn try_read(bytes: &[u8], _ctx: ()) -> byte::Result<(Self, usize)> {
+        let offset = &mut 0;
+
+        let control: SecurityControl = bytes.read(offset)?;
+        let frame_counter = bytes.read_with(offset, LE)?;
+
+        let mut header = AuxiliarySecurityHeader {
+            control,
+            frame_counter,
+            key_identifier: None,
+        };
+
+        // Construct the KeyIdentifier in case the KeyIdentifierMode is not None
+        if control.key_id_mode != KeyIdentifierMode::None {
+            let mut key_identifier = KeyIdentifier {
+                key_source: None,
+                key_index: 0,
+            };
+            match control.key_id_mode {
+                KeyIdentifierMode::KeySource4 => {
+                    key_identifier.key_source = Some(KeySource::Short(bytes.read(offset)?));
+                }
+                KeyIdentifierMode::KeySource8 => {
+                    key_identifier.key_source = Some(KeySource::Long(bytes.read(offset)?));
+                }
+                _ => {}
+            }
+            let key_index = bytes.read(offset)?;
+            key_identifier.key_index = key_index;
+            header.key_identifier = Some(key_identifier);
+        }
+
+        Ok((header, *offset))
+    }
+}
+
+impl TryWrite for AuxiliarySecurityHeader {
+    fn try_write(mut self, bytes: &mut [u8], _ctx: ()) -> byte::Result<usize> {
+        let offset = &mut 0;
+
+        // Set the key id mode to that corresponding to the configured
+        // key identifier in the control field
+        self.control.key_id_mode = match self.key_identifier {
+            Some(key_id) => match key_id.key_source {
+                Some(key_source) => match key_source {
+                    KeySource::Short(_) => KeyIdentifierMode::KeySource4,
+                    KeySource::Long(_) => KeyIdentifierMode::KeySource8,
+                },
+                None => KeyIdentifierMode::KeyIndex,
+            },
+            None => KeyIdentifierMode::None,
+        };
+
+        bytes.write(offset, self.control)?;
+        bytes.write(offset, self.frame_counter)?;
+        match self.key_identifier {
+            Some(key_identifier) => {
+                bytes.write(offset, key_identifier)?;
+            }
+            _ => {}
+        }
+        Ok(*offset)
+    }
+}
+
+/// A key identifier
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+pub struct KeyIdentifier {
+    /// The key source to be used for determining a key from this key identifier (if any)
+    pub key_source: Option<KeySource>,
+    /// The key index to use for determining a key from this key identifier
+    pub key_index: u8,
+}
+
+impl TryWrite for KeyIdentifier {
+    fn try_write(self, bytes: &mut [u8], _ctx: ()) -> byte::Result<usize> {
+        let offset = &mut 0;
+        match self.key_source {
+            Some(source) => match source {
+                KeySource::Short(src) => bytes.write(offset, src)?,
+                KeySource::Long(src) => bytes.write(offset, src)?,
+            },
+            _ => {}
+        }
+
+        bytes.write(offset, self.key_index)?;
+
+        Ok(*offset)
+    }
+}
+
+/// A key source
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+pub enum KeySource {
+    /// A key source that is 4 octets long
+    Short(u32),
+    /// A key source that is 8 octets long
+    Long(u64),
+}
+
+/// A partial key descriptor
+pub struct KeyDescriptor {
+    /// The key contained by this key descriptor
+    pub key: [u8; 16],
+}
